fix: cast row_count to bigint in empty-df fallback of save_volume_metric (#22765)
F.lit(0) defaults to IntegerType, but the normal path produces LongType
via count("*").cast("bigint"). This type mismatch caused Delta to raise
DELTA_FAILED_TO_MERGE_FIELDS when appending to datalake_sst_metrics tables.

Made-with: Cursor

+++ bietlejuice/base/sst/core/observability/metrics.py
@@ -62,7 +62,7 @@ def save_volume_metric(
             .withColumn("source_table", F.lit(table_name))
             .withColumn("environment", F.lit(env))
             .withColumn("layer", F.lit(layer))
-            .withColumn("row_count", F.lit(0))
+            .withColumn("row_count", F.lit(0).cast("bigint"))
             .withColumn("_write_timestamp", F.lit(write_timestamp))
             .select(
                 "metric_category",
